pyproject.toml
[tool.poetry]
name = "varta-svc-agent"
version = "0.1.2"
description = "Agent service for chat + RAG over site data (mock-first)"
authors = ["Varta Team"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
# Pin to versions with wide prebuilt wheel support (avoid 3.14 until SciPy stack is ready)
python = ">=3.11,<3.14"
fastapi = ">=0.115"
uvicorn = {version = ">=0.30", extras = ["standard"]}
pydantic = ">=2.8"
httpx = ">=0.27"
python-dotenv = ">=1.0"
numpy = ">=1.26"
tldextract = ">=5.1"
openai = ">=1.0.0"

# NOTE: No scikit-learn / scipy in base deps to avoid Fortran / PEP517 builds.

[tool.poetry.group.embeddings]
optional = true

[tool.poetry.group.embeddings.dependencies]
# Opt-in heavy deps only when you want real embeddings:
# Ensure your Python has compatible wheels for these versions.
torch = "==2.9.0"
sentence-transformers = ">=3.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.3"
pytest-asyncio = ">=0.23"
black = ">=25.9.0"
ruff = ">=0.14"
mypy = ">=1.10"

[build-system]
requires = ["poetry-core>=1.8"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
src=["src"]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
# F = pyflakes (incl. F401 unused imports)
# I = isort rules (import sorting)
# UP = pyupgrade, B = bugbear, SIM = complexity, ARG = unused args, RUF = extra rules
select = ["E", "F", "I", "UP", "B", "SIM", "ARG", "RUF"]
ignore = []
fixable = ["ALL"]        # allow --fix to apply everything

[tool.ruff.lint.isort]
known-first-party = ["src"]
combine-as-imports = true
force-sort-within-sections = true

[tool.pytest.ini_options]
addopts = "-q"
